requirements.txt
python-telegram-bot[job-queue]==20.7
aiocryptopay==1.0.0
httpx==0.25.1
python-dotenv==1.0.0